fix: setup of PreRinseSpray

+++ app/src/main/java/com/gemini/energy/wordDocGenerator/SorterForWordDocumentGenerator.kt
@@ -794,7 +794,7 @@ class SorterForWordDocumentGenerator {
                             "Pre Rinse Spray",
                             single.energyPowerChange(),
                             costElectricity,
-                            single.age,
+                            single.age.toDouble(),
                             materialCost,
                             implementationCost,
                             savings,
